Centralize bundle artifact projections
