+++ inc/Engine/Bundle/AgentBundleLifecycleProjection.php
@@ -69,7 +69,7 @@ final class AgentBundleLifecycleProjection {
 				'artifact_type' => 'pipeline',
 				'artifact_id'   => $slug,
 				'source_path'   => 'pipelines/' . $slug . '.json',
-				'payload'       => self::pipeline_payload( $pipeline, $slug ),
+				'payload'       => AgentBundleArtifactPayloads::pipeline_payload( $pipeline, $slug ),
 			);
 		}
 
@@ -96,7 +96,7 @@ final class AgentBundleLifecycleProjection {
 				'artifact_type' => 'flow',
 				'artifact_id'   => $slug,
 				'source_path'   => 'flows/' . $slug . '.json',
-				'payload'       => self::flow_payload( $flow, $slug ),
+				'payload'       => AgentBundleArtifactPayloads::flow_payload( $flow, $slug ),
 			);
 		}
 
@@ -149,7 +149,7 @@ final class AgentBundleLifecycleProjection {
 					'artifact_type' => 'pipeline',
 					'artifact_id'   => $id,
 					'source_path'   => (string) ( $record['source_path'] ?? '' ),
-					'payload'       => self::pipeline_payload( $pipeline_by_slug[ $id ], $id ),
+					'payload'       => AgentBundleArtifactPayloads::pipeline_payload( $pipeline_by_slug[ $id ], $id ),
 				);
 			}
 			if ( 'flow' === $type && isset( $flow_by_slug[ $id ] ) ) {
@@ -158,7 +158,7 @@ final class AgentBundleLifecycleProjection {
 					'artifact_type' => 'flow',
 					'artifact_id'   => $id,
 					'source_path'   => (string) ( $record['source_path'] ?? '' ),
-					'payload'       => self::flow_payload( $flow_by_slug[ $id ], $id, $installed_payload ),
+					'payload'       => AgentBundleArtifactPayloads::flow_payload( $flow_by_slug[ $id ], $id, $installed_payload ),
 				);
 			}
 			if ( in_array( $type, self::bundle_file_artifact_types(), true ) ) {
@@ -186,14 +186,14 @@ final class AgentBundleLifecycleProjection {
 		$artifacts = array();
 		$files     = is_array( $bundle['artifact_files'] ?? null ) ? $bundle['artifact_files'] : array();
 
-		foreach ( self::bundle_file_artifact_directories() as $directory => $type ) {
+		foreach ( AgentBundleArtifactDefinitions::file_artifacts() as $directory => $definition ) {
 			foreach ( is_array( $files[ $directory ] ?? null ) ? $files[ $directory ] : array() as $relative_path => $payload ) {
 				$artifact_id = is_array( $payload ) && is_string( $payload['artifact_id'] ?? null )
 					? (string) $payload['artifact_id']
 					: self::artifact_id_from_relative_path( (string) $relative_path );
 
 				$artifacts[] = array(
-					'artifact_type' => $type,
+					'artifact_type' => $definition['artifact_type'],
 					'artifact_id'   => $artifact_id,
 					'source_path'   => $directory . '/' . ltrim( (string) $relative_path, '/' ),
 					'payload'       => $payload,
@@ -204,20 +204,9 @@ final class AgentBundleLifecycleProjection {
 		return $artifacts;
 	}
 
-	/** @return array<string,string> */
-	private static function bundle_file_artifact_directories(): array {
-		return array(
-			BundleSchema::PROMPTS_DIR       => 'prompt',
-			BundleSchema::RUBRICS_DIR       => 'rubric',
-			BundleSchema::TOOL_POLICIES_DIR => 'tool_policy',
-			BundleSchema::AUTH_REFS_DIR     => 'auth_ref',
-			BundleSchema::SEED_QUEUES_DIR   => 'seed_queue',
-		);
-	}
-
 	/** @return string[] */
 	private static function bundle_file_artifact_types(): array {
-		return array_values( self::bundle_file_artifact_directories() );
+		return AgentBundleArtifactDefinitions::file_artifact_types();
 	}
 
 	private static function artifact_id_from_relative_path( string $relative_path ): string {
@@ -240,120 +229,6 @@ final class AgentBundleLifecycleProjection {
 		}
 
 		return null;
-	}
-
-	private static function pipeline_payload( array $pipeline, string $portable_slug ): array {
-		return array(
-			'portable_slug'   => $portable_slug,
-			'pipeline_name'   => (string) ( $pipeline['pipeline_name'] ?? '' ),
-			'pipeline_config' => is_array( $pipeline['pipeline_config'] ?? null ) ? $pipeline['pipeline_config'] : array(),
-		);
-	}
-
-	private static function flow_payload( array $flow, string $portable_slug, ?array $installed_payload = null ): array {
-		$scheduling_policy = self::flow_scheduling_policy( is_array( $flow['scheduling_config'] ?? null ) ? $flow['scheduling_config'] : array() );
-
-		return array(
-			'portable_slug'     => $portable_slug,
-			'flow_name'         => (string) ( $flow['flow_name'] ?? '' ),
-			'flow_config'       => self::flow_config_without_runtime_queues( is_array( $flow['flow_config'] ?? null ) ? $flow['flow_config'] : array() ),
-			'scheduling_policy' => $scheduling_policy,
-			'queue_policy'      => 'create_seed_upgrade_preserve_existing',
-			'runtime_overlays'  => self::flow_runtime_overlays( $flow, $installed_payload ),
-		);
-	}
-
-	private static function flow_scheduling_policy( array $config ): string {
-		$interval = (string) ( $config['interval'] ?? 'manual' );
-		$enabled  = array_key_exists( 'enabled', $config ) ? false !== $config['enabled'] : 'manual' !== $interval;
-
-		if ( 'manual' === $interval || ! $enabled ) {
-			return 'create_paused_upgrade_preserve_existing';
-		}
-
-		return 'create_bundle_schedule_upgrade_preserve_existing';
-	}
-
-	private static function flow_config_without_runtime_queues( array $flow_config ): array {
-		foreach ( $flow_config as &$step ) {
-			if ( is_array( $step ) ) {
-				unset( $step['prompt_queue'], $step['config_patch_queue'], $step['queue_mode'], $step['_queue_consume_revision'] );
-				unset( $step['handler_config']['max_items'] );
-				if ( empty( $step['handler_config'] ) ) {
-					unset( $step['handler_config'] );
-				}
-				if ( is_array( $step['handler_configs'] ?? null ) ) {
-					foreach ( $step['handler_configs'] as $handler_slug => &$handler_config ) {
-						if ( is_array( $handler_config ) ) {
-							unset( $handler_config['max_items'] );
-							if ( empty( $handler_config ) ) {
-								unset( $step['handler_configs'][ $handler_slug ] );
-							}
-						}
-					}
-					unset( $handler_config );
-					if ( empty( $step['handler_configs'] ) ) {
-						unset( $step['handler_configs'] );
-					}
-				}
-			}
-		}
-		unset( $step );
-
-		return $flow_config;
-	}
-
-	private static function flow_runtime_overlays( array $flow, ?array $installed_payload = null ): array {
-		if ( is_array( $installed_payload['runtime_overlays'] ?? null ) ) {
-			return $installed_payload['runtime_overlays'];
-		}
-
-		$overlays    = array();
-		$flow_config = is_array( $flow['flow_config'] ?? null ) ? $flow['flow_config'] : array();
-		$steps       = array();
-
-		foreach ( $flow_config as $flow_step_id => $step ) {
-			if ( ! is_array( $step ) ) {
-				continue;
-			}
-
-			$step_overlay = array();
-			foreach ( array( 'prompt_queue', 'config_patch_queue', 'queue_mode', '_queue_consume_revision' ) as $field ) {
-				if ( array_key_exists( $field, $step ) ) {
-					$step_overlay[ $field ] = $step[ $field ];
-				}
-			}
-			if ( array_key_exists( 'max_items', $step['handler_config'] ?? array() ) ) {
-				$step_overlay['handler_config'] = array( 'max_items' => $step['handler_config']['max_items'] );
-			}
-			if ( is_array( $step['handler_configs'] ?? null ) ) {
-				foreach ( $step['handler_configs'] as $handler_slug => $handler_config ) {
-					if ( is_array( $handler_config ) && array_key_exists( 'max_items', $handler_config ) ) {
-						$step_overlay['handler_configs'][ (string) $handler_slug ] = array( 'max_items' => $handler_config['max_items'] );
-					}
-				}
-			}
-
-			if ( ! empty( $step_overlay ) ) {
-				ksort( $step_overlay, SORT_STRING );
-				$steps[ (string) $flow_step_id ] = $step_overlay;
-			}
-		}
-
-		if ( ! empty( $steps ) ) {
-			ksort( $steps, SORT_STRING );
-			$overlays['steps'] = $steps;
-		}
-
-		$scheduling = is_array( $flow['scheduling_config'] ?? null ) ? $flow['scheduling_config'] : array();
-		unset( $scheduling['last_run'], $scheduling['next_run'], $scheduling['run_count'], $scheduling['run_artifacts'] );
-		if ( ! empty( $scheduling ) ) {
-			ksort( $scheduling, SORT_STRING );
-			$overlays['scheduling_config'] = $scheduling;
-		}
-
-		ksort( $overlays, SORT_STRING );
-		return $overlays;
 	}
 
 	private function pipelines(): Pipelines {
